ENH: move saxi_predict and saxi_train into shapeaxi folder
